Removing JFatCatClient and dependencies
