+++ biojava-structure/src/main/java/org/biojava/nbio/structure/scop/RemoteScopInstallation.java
@@ -24,8 +24,8 @@
  */
 package org.biojava.nbio.structure.scop;
 
-import org.biojava.nbio.structure.align.client.JFatCatClient;
 import org.biojava.nbio.structure.align.util.URLConnectionTools;
+import org.biojava.nbio.structure.rcsb.ReadUtils;
 import org.biojava.nbio.structure.scop.server.ScopDescriptions;
 import org.biojava.nbio.structure.scop.server.ScopDomains;
 import org.biojava.nbio.structure.scop.server.ScopNodes;
@@ -75,7 +75,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 		try {
 			URL u = new URL(server + "getByCategory?category="+category+"&version="+getScopVersion());
 			InputStream response = URLConnectionTools.getInputStream(u);
-			String xml = JFatCatClient.convertStreamToString(response);
+			String xml = ReadUtils.convertStreamToString(response);
 
 			if(! xml.trim().isEmpty()) {
 				ScopDescriptions container = ScopDescriptions.fromXML(xml);
@@ -93,7 +93,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 		try {
 			URL u = new URL(server + "filterByClassificationId?query="+query+"&version="+getScopVersion());
 			InputStream response = URLConnectionTools.getInputStream(u);
-			String xml = JFatCatClient.convertStreamToString(response);
+			String xml = ReadUtils.convertStreamToString(response);
 
 			if(! xml.trim().isEmpty()) {
 				ScopDescriptions container = ScopDescriptions.fromXML(xml);
@@ -111,7 +111,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 		try {
 			URL u = new URL(server + "getTree?scopId="+domain.getScopId()+"&version="+getScopVersion());
 			InputStream response = URLConnectionTools.getInputStream(u);
-			String xml = JFatCatClient.convertStreamToString(response);
+			String xml = ReadUtils.convertStreamToString(response);
 
 			if(! xml.trim().isEmpty()) {
 				ScopNodes container = ScopNodes.fromXML(xml);
@@ -131,7 +131,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 			URL u = new URL(server + "filterByDomainName?query="+query+"&version="+getScopVersion());
 			//System.out.println(u);
 			InputStream response = URLConnectionTools.getInputStream(u);
-			String xml = JFatCatClient.convertStreamToString(response);
+			String xml = ReadUtils.convertStreamToString(response);
 
 			if(! xml.trim().isEmpty()) {
 				ScopDomains container = ScopDomains.fromXML(xml);
@@ -149,7 +149,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 		try {
 			URL u = new URL(server + "filterByDescription?query="+query+"&version="+getScopVersion());
 			InputStream response = URLConnectionTools.getInputStream(u);
-			String xml = JFatCatClient.convertStreamToString(response);
+			String xml = ReadUtils.convertStreamToString(response);
 
 			if(! xml.trim().isEmpty()) {
 				ScopDescriptions container = ScopDescriptions.fromXML(xml);
@@ -171,7 +171,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 
 			URL u = new URL(server + "getScopDescriptionBySunid?sunid="+sunid+"&version="+getScopVersion());
 			InputStream response = URLConnectionTools.getInputStream(u);
-			String xml = JFatCatClient.convertStreamToString(response);
+			String xml = ReadUtils.convertStreamToString(response);
 
 			if(! xml.trim().isEmpty()) {
 				desc = XMLUtil.getScopDescriptionFromXML(xml);
@@ -189,7 +189,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 		try {
 			URL u = new URL(server + "getDomainsForPDB?pdbId="+pdbId+"&version="+getScopVersion());
 			InputStream response = URLConnectionTools.getInputStream(u);
-			String xml = JFatCatClient.convertStreamToString(response);
+			String xml = ReadUtils.convertStreamToString(response);
 
 			if( !xml.trim().isEmpty()) {
 				ScopDomains container = ScopDomains.fromXML(xml);
@@ -206,7 +206,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 		scopId = scopId.trim();
 		URL u = new URL(server + "getDomainByScopID?scopId="+scopId+"&version="+getScopVersion());
 		InputStream response = URLConnectionTools.getInputStream(u);
-		String xml = JFatCatClient.convertStreamToString(response);
+		String xml = ReadUtils.convertStreamToString(response);
 
 		if( !xml.trim().isEmpty()) {
 			return XMLUtil.getScopDomainFromXML(xml);
@@ -229,7 +229,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 		try {
 			URL u = new URL(server + "getScopNode?sunid="+sunid+"&version="+getScopVersion());
 			InputStream response = URLConnectionTools.getInputStream(u);
-			String xml = JFatCatClient.convertStreamToString(response);
+			String xml = ReadUtils.convertStreamToString(response);
 
 			if( !xml.trim().isEmpty()) {
 				desc = XMLUtil.getScopNodeFromXML(xml);
@@ -247,7 +247,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 			try {
 				URL u = new URL(server + "getScopVersion");
 				InputStream response = URLConnectionTools.getInputStream(u);
-				version = JFatCatClient.convertStreamToString(response);
+				version = ReadUtils.convertStreamToString(response);
 				if( version != null)
 					version = version.trim();
 
@@ -269,7 +269,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 		try {
 			URL u = new URL(server + "getScopDomainsBySunid?sunid="+sunid+"&version="+getScopVersion());
 			InputStream response = URLConnectionTools.getInputStream(u);
-			String xml = JFatCatClient.convertStreamToString(response);
+			String xml = ReadUtils.convertStreamToString(response);
 
 			if( !xml.trim().isEmpty()) {
 				ScopDomains container = ScopDomains.fromXML(xml);
@@ -288,7 +288,7 @@ public class RemoteScopInstallation implements ScopDatabase {
 		try {
 			URL u = new URL(server + "getComments?sunid="+sunid+"&version="+getScopVersion());
 			InputStream response = URLConnectionTools.getInputStream(u);
-			String xml = JFatCatClient.convertStreamToString(response);
+			String xml = ReadUtils.convertStreamToString(response);
 
 			if( !xml.trim().isEmpty()) {
 				results = XMLUtil.getCommentsFromXML(xml);
